+++ packages/mona-clients/mona-client-web/src/apis/index.ts
@@ -1,4 +1,4 @@
 import * as AllApi from './api';
 import { BaseApis } from '@bytedance/mona';
 
-export const Mona: BaseApis = AllApi;
+export const Mona = AllApi as BaseApis;
