+++ packages/mona/src/index.ts
@@ -5,6 +5,5 @@ export { default as BaseApis } from './BaseApis';
 export * from './BaseApis';
 export { default as BaseComponents } from './BaseComponents';
 export * from './BaseComponents';
-export * from './EventType';
 export * from './types';
 export * from './lifecycle';
